# perf/contract/after-schema/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateDatabaseConnectionRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "CreateDatabaseConnectionRequest",
        "type": "object"
      },
...
      "DatabaseConnectionSource": {
        "properties": {
          "connection_id": {
            "title": "Connection Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ],
            "title": "Created At"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "table_count": {
            "title": "Table Count",
            "type": "integer"
          },
          "tables": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DbTableInfo"
            },
            "title": "Tables",
            "type": "array"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "connection_id",
          "label",
          "url",
          "status",
          "table_count",
          "created_at"
        ],
        "title": "DatabaseConnectionSource",
        "type": "object"
      },
      "DatabaseConnectionsResponse": {
        "properties": {
          "connections": {
            "items": {
              "$ref": "#/components/schemas/DatabaseConnectionSource"
            },
            "title": "Connections",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "connections",
          "count"
        ],
        "title": "DatabaseConnectionsResponse",
        "type": "object"
      },
      "DbColumnInfo": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nullable": {
            "default": true,
            "title": "Nullable",
            "type": "boolean"
          },
          "primary_key": {
            "default": false,
            "title": "Primary Key",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "DbColumnInfo",
        "type": "object"
      },
      "DbTableInfo": {
        "properties": {
          "columns": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DbColumnInfo"
            },
            "title": "Columns",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          }
        },
        "required": [
          "name"
        ],
        "title": "DbTableInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SetDatabaseConnectionActiveRequest": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "connection_id": {
            "title": "Connection Id",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "active"
        ],
        "title": "SetDatabaseConnectionActiveRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for querying PDF documents with natural language",
    "title": "PDF QA API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/database-connections": {
      "get": {
        "operationId": "list_database_connections_api_v1_database_connections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Database Connections"
      },
      "post": {
        "operationId": "create_database_connection_api_v1_database_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Database Connection"
      }
    },
    "/api/v1/database-connections/activate": {
      "post": {
        "operationId": "set_database_connection_active_api_v1_database_connections_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDatabaseConnectionActiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Database Connection Active"
      }
    },
    "/api/v1/database-connections/{connection_id}": {
      "delete": {
        "operationId": "delete_database_connection_api_v1_database_connections__connection_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Database Connection"
      }
    },
    "/api/v1/database-connections/{connection_id}/tables": {
      "get": {
        "operationId": "refresh_database_connection_tables_api_v1_database_connections__connection_id__tables_get",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseConnectionSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Refresh Database Connection Tables"
      }
    },
...
  }
}